+++ src/views/ProofAddMultiple.vue
@@ -55,9 +55,9 @@
                 color="primary"
                 :block="!$vuetify.display.smAndUp"
                 prepend-icon="mdi-account-circle"
-                @click="goToDashboard"
+                @click="goToUserDashboard"
               >
-                {{ $t('Common.Dashboard') }}
+                {{ $t('Common.MyDashboard') }}
               </v-btn>
             </v-col>
           </v-row>
@@ -101,7 +101,7 @@ export default {
     reloadPage() {
       window.location.reload()
     },
-    goToDashboard() {
+    goToUserDashboard() {
       this.$router.push({ path: '/dashboard', query: { proofSingleSuccess: 'true' } })
     }
   }